Started work on the step parser

+++ src/DinnerScribe.Core/RecipeModel/Types.fs
@@ -2,4 +2,4 @@ module DinnerScribe.Core.RecipeModel.Types
 
 type Title = { Title: string }
 
-type Step = { Number: uint64; Direction: string }
+type Step = { Number: uint; Direction: string }
